Story 2321. Reworked based on Brian's review comments.
